moved the MetricsModule class inside the vgs.metrics folder
