+++ larky/src/test/resources/META-INF/services/com.verygood.security.larky.modules.vgs.metrics.spi.LarkyMetrics
@@ -1,1 +1,1 @@
-com.verygood.security.larky.modules.vgs.metrics.defaults.DefaultMetrics
+com.verygood.security.larky.modules.vgs.metrics.DefaultMetrics
